laundry-backend/requirements.txt
Django>=4.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.0.0
django-cors-headers>=3.13.0
python-dotenv>=0.19.0
Pillow>=8.0.0
requests>=2.25.1

